Defer some analysis to the code-fixer.
