+++ WTG.Analyzers/Analyzers/CodeContracts/CodeContractsAnalyzer.cs
@@ -1,5 +1,4 @@
 using System.Collections.Immutable;
-using System.Threading;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -14,11 +13,7 @@ namespace WTG.Analyzers
 	{
 		public const string PropertyProposedFix = "FIX";
 
-		public const string FixUnavailable = "U";
 		public const string FixDelete = "D";
-		public const string FixGenericRequires = "RG";
-		public const string FixRequiresNonNull = "RN";
-		public const string FixRequiresNonEmptyString = "RS";
 		public const string FixRequires = "R";
 
 		public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics { get; } = ImmutableArray.Create(
@@ -69,7 +64,7 @@ namespace WTG.Analyzers
 					return;
 			}
 
-			context.ReportDiagnostic(FixDeleteDiagnostic(location));
+			context.ReportDiagnostic(Diagnostic.Create(Rules.DoNotUseCodeContractsRule, location, FixDeleteProperties));
 		}
 
 		void AnalyzeInvoke(SyntaxNodeAnalysisContext context)
@@ -82,7 +77,7 @@ namespace WTG.Analyzers
 				return;
 			}
 
-			bool isRequires;
+			ImmutableDictionary<string, string> properties;
 
 			switch (name.Identifier.Text)
 			{
@@ -91,11 +86,12 @@ namespace WTG.Analyzers
 				case nameof(SDC.Contract.EndContractBlock):
 				case nameof(SDC.Contract.Ensures):
 				case nameof(SDC.Contract.EnsuresOnThrow):
-					isRequires = false;
+					properties = FixDeleteProperties;
 					break;
 
 				case nameof(SDC.Contract.Requires):
-					isRequires = true;
+					// Only Contract.Requires needs special handling, all the rest are simply deleted.
+					properties = FixRequiresProperties;
 					break;
 
 				default:
@@ -109,208 +105,11 @@ namespace WTG.Analyzers
 				return;
 			}
 
-			if (!invoke.Parent.IsKind(SyntaxKind.ExpressionStatement))
-			{
-				// All these methods return void, so if the parent is not an ExpressionStatement,
-				// then something crazy is happening and we can't provide a meaningful auto-fix.
-				context.ReportDiagnostic(FixUnavailableDiagnostic(invoke.GetLocation()));
-			}
-			else
-			{
-				var statement = (ExpressionStatementSyntax)invoke.Parent;
+			var location = invoke.Parent.IsKind(SyntaxKind.ExpressionStatement)
+				? ((ExpressionStatementSyntax)invoke.Parent).GetLocation()
+				: invoke.GetLocation();
 
-				if (!isRequires)
-				{
-					// Only Contract.Requires needs special handling, all the rest are simply deleted.
-					context.ReportDiagnostic(FixDeleteDiagnostic(statement.GetLocation()));
-				}
-				else if (symbol.IsGenericMethod)
-				{
-					context.ReportDiagnostic(FixGenericRequiresDiagnostic(statement.GetLocation()));
-				}
-				else if (IsInPrivateMember(context.SemanticModel, statement, context.CancellationToken))
-				{
-					context.ReportDiagnostic(FixDeleteDiagnostic(statement.GetLocation()));
-				}
-				else if (IsNullArgumentCheck(context.SemanticModel, invoke, out var identifierLocation, context.CancellationToken))
-				{
-					context.ReportDiagnostic(FixRequiresNonNullDiagnostic(statement.GetLocation(), identifierLocation));
-				}
-				else if (IsNonEmptyStringArgumentCheck(context.SemanticModel, invoke, out identifierLocation, context.CancellationToken))
-				{
-					context.ReportDiagnostic(FixRequiresNonEmptyStringDiagnostic(statement.GetLocation(), identifierLocation));
-				}
-				else if (RequiresParameter(context.SemanticModel, invoke, out identifierLocation, context.CancellationToken))
-				{
-					context.ReportDiagnostic(FixRequiresDiagnostic(statement.GetLocation(), identifierLocation));
-				}
-				else
-				{
-					context.ReportDiagnostic(FixUnavailableDiagnostic(statement.GetLocation()));
-				}
-			}
-		}
-
-		static bool IsNullArgumentCheck(SemanticModel semanticModel, InvocationExpressionSyntax invoke, out Location identifierLocation, CancellationToken cancellationToken)
-		{
-			var arguments = invoke.ArgumentList.Arguments;
-
-			if (arguments.Count == 0)
-			{
-				identifierLocation = null;
-				return false;
-			}
-
-			var comparand = GetComparand(arguments[0].Expression);
-
-			if (comparand == null)
-			{
-				identifierLocation = null;
-				return false;
-			}
-
-			var symbol = semanticModel.GetSymbolInfo(comparand, cancellationToken).Symbol;
-
-			if (symbol == null || symbol.Kind != SymbolKind.Parameter)
-			{
-				identifierLocation = null;
-				return false;
-			}
-
-			identifierLocation = comparand.GetLocation();
-			return true;
-
-			ExpressionSyntax GetComparand(ExpressionSyntax condition)
-			{
-				if (!condition.IsKind(SyntaxKind.NotEqualsExpression))
-				{
-					return null;
-				}
-
-				var b = (BinaryExpressionSyntax)condition;
-
-				if (b.Right.IsKind(SyntaxKind.NullLiteralExpression))
-				{
-					return b.Left;
-				}
-				else if (b.Left.IsKind(SyntaxKind.NullLiteralExpression))
-				{
-					return b.Right;
-				}
-				else
-				{
-					return null;
-				}
-			}
-		}
-
-		static bool IsNonEmptyStringArgumentCheck(SemanticModel semanticModel, InvocationExpressionSyntax invoke, out Location identifierLocation, CancellationToken cancellationToken)
-		{
-			var arguments = invoke.ArgumentList.Arguments;
-
-			if (arguments.Count == 0)
-			{
-				identifierLocation = null;
-				return false;
-			}
-
-			var expression = arguments[0].Expression;
-
-			if (!expression.IsKind(SyntaxKind.LogicalNotExpression))
-			{
-				identifierLocation = null;
-				return false;
-			}
-
-			expression = ((PrefixUnaryExpressionSyntax)expression).Operand;
-
-			if (!expression.IsKind(SyntaxKind.InvocationExpression))
-			{
-				identifierLocation = null;
-				return false;
-			}
-
-			var checkInvoke = (InvocationExpressionSyntax)expression;
-			var checkArguments = checkInvoke.ArgumentList.Arguments;
-
-			if (ExpressionHelper.GetMethodName(checkInvoke).Identifier.Text != nameof(string.IsNullOrEmpty) ||
-				checkArguments.Count != 1)
-			{
-				identifierLocation = null;
-				return false;
-			}
-
-			var checkMethodSymbol = (IMethodSymbol)semanticModel.GetSymbolInfo(checkInvoke, cancellationToken).Symbol;
-
-			if (checkMethodSymbol == null || checkMethodSymbol.ContainingType.SpecialType != SpecialType.System_String)
-			{
-				identifierLocation = null;
-				return false;
-			}
-
-			expression = checkArguments[0].Expression;
-			var paramSymbol = semanticModel.GetSymbolInfo(expression, cancellationToken).Symbol;
-
-			if (paramSymbol == null || paramSymbol.Kind != SymbolKind.Parameter)
-			{
-				identifierLocation = null;
-				return false;
-			}
-
-			identifierLocation = expression.GetLocation();
-			return true;
-		}
-
-		static bool RequiresParameter(SemanticModel model, InvocationExpressionSyntax invoke, out Location identifierLocation, CancellationToken cancellationToken)
-		{
-			var locator = new ParameterLocator(model, cancellationToken);
-			invoke.ArgumentList.Accept(locator);
-			identifierLocation = locator.ParameterLocation;
-			return identifierLocation != null;
-		}
-
-		static bool IsInPrivateMember(SemanticModel model, SyntaxNode node, CancellationToken cancellationToken)
-		{
-			while (node != null)
-			{
-				switch (node.Kind())
-				{
-					case SyntaxKind.MethodDeclaration:
-						var methodDecl = (MethodDeclarationSyntax)node;
-						return methodDecl.ExplicitInterfaceSpecifier == null && IsPrivate(model.GetDeclaredSymbol(methodDecl));
-
-					case SyntaxKind.ConstructorDeclaration:
-						return IsPrivate(model.GetDeclaredSymbol((ConstructorDeclarationSyntax)node, cancellationToken));
-
-					case SyntaxKind.DestructorDeclaration:
-						return IsPrivate(model.GetDeclaredSymbol((DestructorDeclarationSyntax)node, cancellationToken));
-
-					case SyntaxKind.AddAccessorDeclaration:
-					case SyntaxKind.RemoveAccessorDeclaration:
-						var eventAccessorDecl = (AccessorDeclarationSyntax)node;
-						var eventDecl = (EventDeclarationSyntax)node.Parent;
-						return eventDecl?.ExplicitInterfaceSpecifier != null && IsPrivate(model.GetDeclaredSymbol(eventAccessorDecl, cancellationToken));
-
-					case SyntaxKind.GetAccessorDeclaration:
-					case SyntaxKind.SetAccessorDeclaration:
-						var propertyAccessorDecl = (AccessorDeclarationSyntax)node;
-						var propertyDecl = (PropertyDeclarationSyntax)node.Parent;
-						return propertyDecl?.ExplicitInterfaceSpecifier != null && IsPrivate(model.GetDeclaredSymbol(propertyAccessorDecl, cancellationToken));
-
-					case SyntaxKind.NamespaceDeclaration:
-					case SyntaxKind.CompilationUnit:
-					case SyntaxKind.ClassDeclaration:
-					case SyntaxKind.StructDeclaration:
-					case SyntaxKind.UnknownAccessorDeclaration:
-						return false;
-				}
-
-				node = node.Parent;
-			}
-
-			return false;
-
-			bool IsPrivate(ISymbol symbol) => symbol != null && symbol.DeclaredAccessibility == Accessibility.Private;
+			context.ReportDiagnostic(Diagnostic.Create(Rules.DoNotUseCodeContractsRule, location, properties));
 		}
 
 		static Location GetAttributedMemberLocation(AttributeSyntax attribute, SyntaxKind expectedKind)
@@ -330,77 +129,7 @@ namespace WTG.Analyzers
 			return AttributeUtils.GetLocation(attribute);
 		}
 
-		static Diagnostic FixUnavailableDiagnostic(Location location)
-			=> Diagnostic.Create(Rules.DoNotUseCodeContractsRule, location, FixUnavailableProperties);
-
-		static Diagnostic FixDeleteDiagnostic(Location location)
-			=> Diagnostic.Create(Rules.DoNotUseCodeContractsRule, location, FixDeleteProperties);
-
-		static Diagnostic FixGenericRequiresDiagnostic(Location location)
-			=> Diagnostic.Create(Rules.DoNotUseCodeContractsRule, location, FixGenericRequiresProperties);
-
-		static Diagnostic FixRequiresNonNullDiagnostic(Location location, Location identifierLocation)
-			=> Diagnostic.Create(Rules.DoNotUseCodeContractsRule, location, new[] { identifierLocation }, FixRequiresNonNullProperties);
-
-		static Diagnostic FixRequiresNonEmptyStringDiagnostic(Location location, Location identifierLocation)
-			=> Diagnostic.Create(Rules.DoNotUseCodeContractsRule, location, new[] { identifierLocation }, FixRequiresNonEmptyStringProperties);
-
-		static Diagnostic FixRequiresDiagnostic(Location location, Location identifierLocation)
-			=> Diagnostic.Create(Rules.DoNotUseCodeContractsRule, location, new[] { identifierLocation }, FixRequiresProperties);
-
-		static readonly ImmutableDictionary<string, string> FixUnavailableProperties = ImmutableDictionary<string, string>.Empty.Add(PropertyProposedFix, FixUnavailable);
 		static readonly ImmutableDictionary<string, string> FixDeleteProperties = ImmutableDictionary<string, string>.Empty.Add(PropertyProposedFix, FixDelete);
-		static readonly ImmutableDictionary<string, string> FixGenericRequiresProperties = ImmutableDictionary<string, string>.Empty.Add(PropertyProposedFix, FixGenericRequires);
-		static readonly ImmutableDictionary<string, string> FixRequiresNonNullProperties = ImmutableDictionary<string, string>.Empty.Add(PropertyProposedFix, FixRequiresNonNull);
-		static readonly ImmutableDictionary<string, string> FixRequiresNonEmptyStringProperties = ImmutableDictionary<string, string>.Empty.Add(PropertyProposedFix, FixRequiresNonEmptyString);
 		static readonly ImmutableDictionary<string, string> FixRequiresProperties = ImmutableDictionary<string, string>.Empty.Add(PropertyProposedFix, FixRequires);
-
-		sealed class ParameterLocator : CSharpSyntaxWalker
-		{
-			public ParameterLocator(SemanticModel model, CancellationToken cancellationToken)
-			{
-				this.model = model;
-				this.cancellationToken = cancellationToken;
-			}
-
-			public Location ParameterLocation { get; private set; }
-
-			public override void Visit(SyntaxNode node)
-			{
-				if (ParameterLocation == null)
-				{
-					base.Visit(node);
-				}
-			}
-
-			public override void VisitIdentifierName(IdentifierNameSyntax node)
-			{
-				var symbol = model.GetSymbolInfo(node, cancellationToken).Symbol;
-
-				if (symbol != null && symbol.Kind == SymbolKind.Parameter)
-				{
-					ParameterLocation = node.GetLocation();
-				}
-			}
-
-			public override void VisitParenthesizedLambdaExpression(ParenthesizedLambdaExpressionSyntax node)
-			{
-			}
-
-			public override void VisitSimpleLambdaExpression(SimpleLambdaExpressionSyntax node)
-			{
-			}
-
-			public override void VisitAnonymousMethodExpression(AnonymousMethodExpressionSyntax node)
-			{
-			}
-
-			public override void VisitLocalFunctionStatement(LocalFunctionStatementSyntax node)
-			{
-			}
-
-			readonly SemanticModel model;
-			readonly CancellationToken cancellationToken;
-		}
 	}
 }
